+++ PersonalTools/src/main/java/com/robbie/personaltools/infra/databases/dao/user/AccountDao.java
@@ -1,6 +1,6 @@
-package com.robbie.personaltools.infra.databases.dao.cheatmeal;
+package com.robbie.personaltools.infra.databases.dao.user;
 
 import com.robbie.personaltools.infra.databases.entity.user.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberDao extends JpaRepository<Account, String> {}
+public interface AccountDao extends JpaRepository<Account, String> {}
