+++ PersonalTools/src/main/java/com/robbie/personaltools/middle/infrastructure/persistence/AccountPersistence.java
@@ -1,6 +1,6 @@
 package com.robbie.personaltools.middle.infrastructure.persistence;
 
-import com.robbie.personaltools.infra.databases.dao.cheatmeal.MemberDao;
+import com.robbie.personaltools.infra.databases.dao.user.AccountDao;
 import com.robbie.personaltools.infra.databases.entity.user.Account;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
@@ -8,10 +8,10 @@ import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class MemberPersistence {
-  private final MemberDao memberDao;
+public class AccountPersistence {
+  private final AccountDao accountDao;
 
   public Optional<Account> findByUserId(String userId) {
-    return this.memberDao.findById(userId);
+    return this.accountDao.findById(userId);
   }
 }
